Two basic modifications:  - Simplified the way in which the error_log works.  - Added a liokinds modules for kind declarations.
Also, incorporated the concept of module "tiers" (the idea will be
that eventually the compilation will do first all tier 0 modules
which don't depend on anything else, then tier 1 modules which only
depend on tier 0 modules, etc). If it doesn't stick, it is just a
comment to be removed in the future.